requirements.txt
streamlit
pdfplumber
python-docx
spacy
nltk
openai
fpdf
